tests/wpt/web-platform-tests/svg/shapes/scripted/stroke-dashes-hit-at-high-scale.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" id="svg" width="600" height="600" viewBox="0 0 36 36"><title>Strokes w/dashes are properly hit-tested, even at large scale factors</title><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><metadata><h:link href="https://svgwg.org/svg2-draft/shapes.html#CircleElement" rel="help"/><h:link href="https://svgwg.org/svg2-draft/painting.html#StrokeProperties" rel="help"/></metadata><circle id="circle" cx="6" cy="10" r="5" fill="none" stroke="#00f" stroke-dasharray="10 21.416" stroke-width="1"/><script><![CDATA[
    test(function() {
      let svg = document.getElementById("svg");
      let circle = document.getElementById("circle");
      let hitTest = function(x, y) {
        return document.elementFromPoint(
                x * svg.width.baseVal.value / svg.viewBox.baseVal.width,
                y * svg.height.baseVal.value / svg.viewBox.baseVal.height);
      }
      assert_equals(hitTest(11, 10), circle, "hit-test the beginning of the dash (t=0)");
      assert_equals(hitTest(8.70, 14.21), circle, "hit-test the middle of the dash (t=5)");
      assert_equals(hitTest(4.10, 14.63), circle, "hit-test the end of the dash (t=9.8)");
      assert_equals(hitTest(3.74, 14.46), svg, "hit-test past the end of the dash (t=10.2)");
    });
  ]]></script></svg>
